New Packaging Project
Instead of saving, compressing and uploading the image files in
each module, just save the tag for the image file in a text file
attached to the build. That way, the packaging module can do the
save and compress, delaying the need to do so and having to do it
just once.

Moved runa2kd.md from adept-e2e repo to here close to the script.
Added the md file to the utilities tar.
Added the tag for the docker image to a txt file and attached it to
the build.

Moved runkb.md from adept-e2e repo to here close to the script.
Added the md file to the utilities tar
Added the tag for the docker image to a txt file and attached it to
the build.